Convert hero block to MVC field model
