+++ wp-content/plugins/core/src/Blocks/Blocks_Definer.php
@@ -11,7 +11,6 @@ use Tribe\Project\Blocks\Types\Content_Columns\Content_Columns;
 use Tribe\Project\Blocks\Types\Content_Loop\Content_Loop;
 use Tribe\Project\Blocks\Types\Gallery_Grid\Gallery_Grid;
 use Tribe\Project\Blocks\Types\Gallery_Slider\Gallery_Slider;
-use Tribe\Project\Blocks\Types\Hero\Hero;
 use Tribe\Project\Blocks\Types\Interstitial\Interstitial;
 use Tribe\Project\Blocks\Types\Lead_Form\Lead_Form;
 use Tribe\Project\Blocks\Types\Links\Links;
@@ -22,13 +21,15 @@ use Tribe\Project\Blocks\Types\Section_Nav\Section_Nav;
 use Tribe\Project\Blocks\Types\Spacer\Spacer;
 use Tribe\Project\Blocks\Types\Stats\Stats;
 use Tribe\Project\Blocks\Types\Tabs\Tabs;
+use Tribe\Project\Templates\Components\blocks\hero\src\Hero;
 use Tribe\Project\Templates\Components\blocks\icon_grid\src\Icon_Grid;
 
 class Blocks_Definer implements Definer_Interface {
 
-	public const DENY_LIST = 'blocks.deny_list';
-	public const STYLES    = 'blocks.style_overrides';
-	public const TYPES     = 'blocks.types';
+	public const DENY_LIST  = 'blocks.deny_list';
+	public const STYLES     = 'blocks.style_overrides';
+	public const TYPES      = 'blocks.types';
+	public const BLOCK_PATH = 'blocks.block_path';
 
 	public function define(): array {
 		return [
@@ -117,7 +118,9 @@ class Blocks_Definer implements Definer_Interface {
 			Block_Deny_List::class    => DI\autowire()->constructor( DI\get( self::DENY_LIST ) ),
 
 			// TODO: likely to update this path out of components, since a block isn't technically a component
-			Block_View_Factory::class => DI\autowire()->constructorParameter( 'block_path', 'components/blocks/%1$s/%1$s' ),
+			self::BLOCK_PATH          => DI\string( 'components/blocks/%1$s/%1$s' ),
+			Block_View_Factory::class => DI\autowire()->constructorParameter( 'block_path', DI\get( self::BLOCK_PATH ) ),
+			Block_Component::class    => DI\autowire()->constructorParameter( 'block_path', DI\get( self::BLOCK_PATH ) ),
 		];
 	}
