--- wp-content/themes/core/components/blocks/hero/Hero_Block_Controller.php
@@ -1,222 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Tribe\Project\Templates\Components\blocks\hero;
-
-use Tribe\Libs\Field_Models\Models\Cta;
-use Tribe\Libs\Field_Models\Models\Image;
-use Tribe\Libs\Utils\Markup_Utils;
-use Tribe\Project\Blocks\Types\Hero\Hero as Hero_Block;
-use Tribe\Project\Templates\Components\Abstract_Controller;
-use Tribe\Project\Templates\Components\container\Container_Controller;
-use Tribe\Project\Templates\Components\content_block\Content_Block_Controller;
-use Tribe\Project\Templates\Components\Deferred_Component;
-use Tribe\Project\Templates\Components\image\Image_Controller;
-use Tribe\Project\Templates\Components\link\Link_Controller;
-use Tribe\Project\Templates\Components\text\Text_Controller;
-use Tribe\Project\Theme\Config\Image_Sizes;
-
-class Hero_Block_Controller extends Abstract_Controller {
-
-	public const ATTRS             = 'attrs';
-	public const CLASSES           = 'classes';
-	public const CONTAINER_CLASSES = 'container_classes';
-	public const CONTENT_CLASSES   = 'content_classes';
-	public const CTA               = 'cta';
-	public const DESCRIPTION       = 'description';
-	public const LAYOUT            = 'layout';
-	public const LEADIN            = 'leadin';
-	public const MEDIA             = 'media';
-	public const MEDIA_CLASSES     = 'media_classes';
-	public const TITLE             = 'title';
-
-	/**
-	 * @var string[]
-	 */
-	private array $attrs;
-
-	/**
-	 * @var string[]
-	 */
-	private array $classes;
-
-	/**
-	 * @var string[]
-	 */
-	private array $container_classes;
-
-	/**
-	 * @var string[]
-	 */
-	private array $content_classes;
-
-	/**
-	 * @var string[]
-	 */
-	private array $media_classes;
-	private Cta $cta;
-	private Image $media;
-	private string $description;
-	private string $layout;
-	private string $leadin;
-	private string $title;
-
-	/**
-	 * @param array $args
-	 */
-	public function __construct( array $args = [] ) {
-		$args = $this->parse_args( $args );
-
-		$this->attrs             = (array) $args[ self::ATTRS ];
-		$this->classes           = (array) $args[ self::CLASSES ];
-		$this->container_classes = (array) $args[ self::CONTAINER_CLASSES ];
-		$this->content_classes   = (array) $args[ self::CONTENT_CLASSES ];
-		$this->cta               = $args[ self::CTA ];
-		$this->description       = (string) $args[ self::DESCRIPTION ];
-		$this->layout            = (string) $args[ self::LAYOUT ];
-		$this->leadin            = (string) $args[ self::LEADIN ];
-		$this->media             = $args[ self::MEDIA ];
-		$this->media_classes     = (array) $args[ self::MEDIA_CLASSES ];
-		$this->title             = (string) $args[ self::TITLE ];
-	}
-
-	public function get_classes(): string {
-		$this->classes[] = 'c-block--layout-' . $this->layout;
-
-		return Markup_Utils::class_attribute( $this->classes );
-	}
-
-	public function get_attrs(): string {
-		return Markup_Utils::concat_attrs( $this->attrs );
-	}
-
-	public function get_container_classes(): string {
-		return Markup_Utils::class_attribute( $this->container_classes );
-	}
-
-	public function get_media_classes(): string {
-		return Markup_Utils::class_attribute( $this->media_classes );
-	}
-
-	public function get_content_classes(): string {
-		return Markup_Utils::class_attribute( $this->content_classes );
-	}
-
-	public function get_content_args(): array {
-		if ( empty( $this->title ) && empty( $this->description ) ) {
-			return [];
-		}
-
-		return [
-			Content_Block_Controller::TAG     => 'header',
-			Content_Block_Controller::LEADIN  => $this->get_leadin(),
-			Content_Block_Controller::TITLE   => $this->get_title(),
-			Content_Block_Controller::CONTENT => $this->get_content(),
-			Content_Block_Controller::CTA     => $this->get_cta(),
-			Content_Block_Controller::LAYOUT  => $this->layout,
-			Content_Block_Controller::CLASSES => [
-				'c-block__content-block',
-				'c-block__header',
-				'b-hero__content-container',
-				't-theme--light',
-			],
-		];
-	}
-
-	public function get_image_args(): array {
-		if ( ! $this->media->id ) {
-			return [];
-		}
-
-		return [
-			Image_Controller::IMG_ID       => $this->media->id,
-			Image_Controller::AS_BG        => true,
-			Image_Controller::AUTO_SHIM    => false,
-			Image_Controller::USE_LAZYLOAD => true,
-			Image_Controller::WRAPPER_TAG  => 'div',
-			Image_Controller::CLASSES      => [ 'b-hero__figure', 'c-image--bg', 'c-image--overlay' ],
-			Image_Controller::IMG_CLASSES  => [ 'b-hero__img' ],
-			Image_Controller::SRC_SIZE     => Image_Sizes::CORE_FULL,
-			Image_Controller::SRCSET_SIZES => [
-				Image_Sizes::CORE_FULL,
-				Image_Sizes::CORE_MOBILE,
-			],
-		];
-	}
-
-	protected function defaults(): array {
-		return [
-			self::ATTRS             => [],
-			self::CLASSES           => [ 'c-block--full-bleed' ],
-			self::CONTAINER_CLASSES => [],
-			self::CONTENT_CLASSES   => [],
-			self::CTA               => new Cta(),
-			self::DESCRIPTION       => '',
-			self::LAYOUT            => Hero_Block::LAYOUT_LEFT,
-			self::LEADIN            => '',
-			self::MEDIA             => new Image(),
-			self::MEDIA_CLASSES     => [],
-			self::TITLE             => '',
-		];
-	}
-
-	protected function required(): array {
-		return [
-			self::CLASSES           => [ 'c-block', 'b-hero' ],
-			self::CONTAINER_CLASSES => [ 'b-hero__container', 'l-container' ],
-			self::CONTENT_CLASSES   => [ 'b-hero__content' ],
-			self::MEDIA_CLASSES     => [ 'b-hero__media' ],
-		];
-	}
-
-	private function get_leadin(): Deferred_Component {
-		return defer_template_part( 'components/text/text', null, [
-			Text_Controller::CLASSES => [
-				'c-block__leadin',
-				'b-hero__leadin',
-				'h6',
-			],
-			Text_Controller::CONTENT => $this->leadin,
-		] );
-	}
-
-	private function get_title(): Deferred_Component {
-		return defer_template_part( 'components/text/text', null, [
-			Text_Controller::TAG     => 'h2',
-			Text_Controller::CLASSES => [
-				'c-block__title',
-				'b-hero__title',
-				'h3',
-			],
-			Text_Controller::CONTENT => $this->title,
-		] );
-	}
-
-	private function get_content(): Deferred_Component {
-		return defer_template_part( 'components/container/container', null, [
-			Container_Controller::CLASSES => [
-				'c-block__description',
-				'b-hero__description',
-				't-sink',
-				's-sink',
-			],
-			Container_Controller::CONTENT => $this->description,
-		] );
-	}
-
-	private function get_cta(): Deferred_Component {
-		return defer_template_part( 'components/link/link', null, [
-			Link_Controller::URL            => $this->cta->link->url,
-			Link_Controller::CONTENT        => $this->cta->link->title ?: $this->cta->link->url,
-			Link_Controller::TARGET         => $this->cta->link->target,
-			Link_Controller::ADD_ARIA_LABEL => $this->cta->add_aria_label,
-			Link_Controller::ARIA_LABEL     => $this->cta->aria_label,
-			Link_Controller::CLASSES        => [
-				'c-block__cta-link',
-				'a-btn',
-				'a-btn--has-icon-after',
-				'icon-arrow-right',
-			],
-		] );
-	}
-
-}
