+++ wp-content/themes/core/components/blocks/hero/src/Hero.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Tribe\Project\Blocks\Types\Hero;
+namespace Tribe\Project\Templates\Components\blocks\hero\src;
 
 use Tribe\Libs\ACF\Block;
 use Tribe\Libs\ACF\Block_Config;
@@ -33,15 +33,18 @@ class Hero extends Block_Config implements Cta_Field {
 	 */
 	public function add_block(): void {
 		$this->set_block( new Block( self::NAME, [
-			'title'       => __( 'Hero', 'tribe' ),
-			'description' => __( 'Hero block', 'tribe' ),
+			'title'       => esc_html__( 'Hero', 'tribe' ),
+			'description' => esc_html__( 'Hero block', 'tribe' ),
 			'icon'        => '<svg width="28" height="19" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M8.254 5.968H19.85v1.054H8.254V5.968zm1.462 1.57h8.568v1.053H9.716V7.538z" fill="#000"/><path d="M2.82 2.437v14.126H25.2V2.437H2.82zm21.482 13.176H3.866V3.366h20.436v12.247z" fill="#000"/><path d="M10.092 16.15H2.884l10.614-3.242 1.693 1.012 5.433-3.243 4.409 2.623v2.85h-14.94z" fill="#000"/></svg>',
-			'keywords'    => [ __( 'hero', 'tribe' ), __( 'display', 'tribe' ) ],
+			'keywords'    => [ esc_html__( 'hero', 'tribe' ), esc_html__( 'display', 'tribe' ) ],
 			'category'    => 'layout',
 			'supports'    => [
 				'align'  => false,
 				'anchor' => true,
 			],
+			'model'       => Hero_Model::class,
+			'view'        => 'hero',
+			'controller'  => Hero_Block_Controller::class,
 			'example'     => [
 				'attributes' => [
 					'mode' => 'preview',
@@ -97,7 +100,7 @@ class Hero extends Block_Config implements Cta_Field {
 			$this->get_cta_field( self::NAME )
 		);
 
-		$this->add_section( new Field_Section( self::SECTION_BACKGROUND, __( 'Background', 'tribe' ), 'accordion' ) )
+		$this->add_section( new Field_Section( self::SECTION_BACKGROUND, esc_html__( 'Background', 'tribe' ), 'accordion' ) )
 			 ->add_field( new Field( self::NAME . '_' . self::IMAGE, [
 					 'label'         => esc_html__( 'Image', 'tribe' ),
 					 'name'          => self::IMAGE,
@@ -110,7 +113,7 @@ class Hero extends Block_Config implements Cta_Field {
 				 ] )
 			 );
 
-		$this->add_section( new Field_Section( self::SECTION_APPEARANCE, __( 'Appearance', 'tribe' ), 'accordion' ) )
+		$this->add_section( new Field_Section( self::SECTION_APPEARANCE, esc_html__( 'Appearance', 'tribe' ), 'accordion' ) )
 			 ->add_field( new Field( self::NAME . '_' . self::LAYOUT, [
 				 'label'         => __( 'Text Alignment', 'tribe' ),
 				 'type'          => 'button_group',
